requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.4
django-countries==7.3.2
djangorestframework==3.13.1
install==1.3.5
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.2.0
